+++ examples/Example.ShowVersion/Properties/AssemblyInfo.cs
@@ -18,7 +18,7 @@ using System.Runtime.InteropServices;
 [assembly: CLSCompliant(true)]
 [assembly: ComVisible(false)]
 
-[assembly: AssemblyInformationalVersion("0.15")]
+[assembly: AssemblyInformationalVersion("0.16")]
 
 /* ///////////////////////////// end of file //////////////////////////// */