+++ src/CLASP.Core/Properties/AssemblyInfo.cs
@@ -21,9 +21,9 @@ using System.Runtime.InteropServices;
 [assembly: CLSCompliant(true)]
 [assembly: ComVisible(false)]
 
-[assembly: AssemblyVersion("0.15.133.1")]
-[assembly: AssemblyFileVersion("0.15.133.1")]
-[assembly: AssemblyInformationalVersion("0.15")]
+[assembly: AssemblyVersion("0.16.134.0")]
+[assembly: AssemblyFileVersion("0.16.134.0")]
+[assembly: AssemblyInformationalVersion("0.16")]
 
 /* ///////////////////////////// end of file //////////////////////////// */